+++ starbucks/src/main/java/git_kkalnane/backend/starbucks/member/dto/request/UpdatePasswordRequest.java
@@ -1,6 +1,6 @@
 package git_kkalnane.backend.starbucks.member.dto.request;
 
-import git_kkalnane.backend.starbucks.member.validation.annotation.ValidPassword;
+import git_kkalnane.backend.starbucks._global.validation.annotation.ValidPassword;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 /**
